refactor: modularized `mod spec_traits` into a `specialization.rs` file
